pyproject.toml
[project]
name = "claude-personal-helper"
version = "1.0.0"
description = "Telegram bot with translation features powered by Google Gemini"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[webhooks]>=22.6",
    "sqlmodel>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    "litestar>=2.19.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "fluent.runtime>=0.4.0",
    "fluent.syntax>=0.19.0",
    "google-genai>=1.66.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
